+++ sunscreen_runtime/src/error.rs
@@ -123,17 +123,10 @@ pub enum Error {
     ZkpError(#[from] ZkpError),
 
     /**
-     * An error occurred when building a proof or verification
+     * An error occurred from incorrect usage of a builder.
      */
-    #[error("ZKP builder error: {0}")]
-    ZkpBuilderError(Box<String>),
-
-    /**
-     * Error when building a logproof
-     */
-    #[cfg(feature = "sdlp")]
-    #[error(transparent)]
-    LogProofBuilderError(#[from] crate::sdlp::LogProofBuilderError),
+    #[error("Builder error: {0}")]
+    BuilderError(#[from] crate::builder::BuilderError),
 
     /// Error when proving or verifying a linked proof.
     #[cfg(feature = "linkedproofs")]
@@ -163,13 +156,6 @@ impl Error {
      */
     pub fn fhe_type_error(msg: &str) -> Self {
         Self::FheTypeError(Box::new(msg.to_owned()))
-    }
-
-    /**
-     * Create an [`Error::ZkpBuilderError`].
-     */
-    pub fn zkp_builder_error(msg: &str) -> Self {
-        Self::ZkpBuilderError(Box::new(msg.to_owned()))
     }
 
     fn unwrap_argument_mismatch_data(&self) -> &(Vec<Type>, Vec<Type>) {
