+++ sunscreen_runtime/src/runtime.rs
@@ -5,6 +5,8 @@ use merlin::Transcript;
 
 use crate::error::*;
 use crate::metadata::*;
+use crate::ProofBuilder;
+use crate::VerificationBuilder;
 use crate::ZkpProgramInput;
 use crate::{
     run_program_unchecked, serialization::WithContext, Ciphertext, FheProgramInput,
@@ -331,6 +333,7 @@ where
     /**
      * Returns the underlying SEAL context.
      */
+    #[allow(unused)]
     pub(crate) fn context(&self) -> &SealContext {
         match &self.runtime_data.unwrap_fhe().context {
             Context::Seal(seal_ctx) => seal_ctx,
@@ -1023,203 +1026,3 @@ impl<B> ZkpRuntime<B> {
  *   can do both.
  */
 pub type Runtime = GenericRuntime<(), ()>;
-
-/// A builder for creating a proof.
-///
-/// This is offered as a convenience for building the arguments necessary for the
-/// [`prove`][GenericRuntime::prove] function.
-pub struct ProofBuilder<'r, 'p, T: marker::Zkp, B: ZkpBackend> {
-    runtime: &'r GenericRuntime<T, B>,
-    program: &'p CompiledZkpProgram,
-    private_inputs: Vec<ZkpProgramInput>,
-    public_inputs: Vec<ZkpProgramInput>,
-    constant_inputs: Vec<ZkpProgramInput>,
-}
-
-impl<'r, 'p, T: marker::Zkp, B: ZkpBackend> ProofBuilder<'r, 'p, T, B> {
-    /// Create a new `ProofBuilder`. It's typically more convenient to create a proof builder
-    /// via [`runtime.proof_builder()`][GenericRuntime::proof_builder].
-    pub fn new(runtime: &'r GenericRuntime<T, B>, program: &'p CompiledZkpProgram) -> Self
-    where
-        T: marker::Zkp,
-        B: ZkpBackend,
-    {
-        Self {
-            runtime,
-            program,
-            private_inputs: vec![],
-            public_inputs: vec![],
-            constant_inputs: vec![],
-        }
-    }
-
-    /// Add a constant input to the proof builder.
-    pub fn constant_input(mut self, input: impl Into<ZkpProgramInput>) -> Self {
-        self.constant_inputs.push(input.into());
-        self
-    }
-
-    /// Add multiple constant inputs to the proof builder.
-    pub fn constant_inputs<I>(mut self, inputs: I) -> Self
-    where
-        I: IntoIterator<Item = T>,
-        ZkpProgramInput: From<T>,
-    {
-        self.constant_inputs
-            .extend(inputs.into_iter().map(ZkpProgramInput::from));
-        self
-    }
-
-    /// Add a public input to the proof builder.
-    pub fn public_input(mut self, input: impl Into<ZkpProgramInput>) -> Self {
-        self.public_inputs.push(input.into());
-        self
-    }
-
-    /// Add multiple public inputs to the proof builder.
-    pub fn public_inputs<I>(mut self, inputs: I) -> Self
-    where
-        I: IntoIterator<Item = T>,
-        ZkpProgramInput: From<T>,
-    {
-        self.public_inputs
-            .extend(inputs.into_iter().map(ZkpProgramInput::from));
-        self
-    }
-
-    /// Add a private input to the proof builder.
-    pub fn private_input(mut self, input: impl Into<ZkpProgramInput>) -> Self {
-        self.private_inputs.push(input.into());
-        self
-    }
-
-    /// Add multiple private inputs to the proof builder.
-    pub fn private_inputs<I>(mut self, inputs: I) -> Self
-    where
-        I: IntoIterator<Item = T>,
-        ZkpProgramInput: From<T>,
-    {
-        self.private_inputs
-            .extend(inputs.into_iter().map(ZkpProgramInput::from));
-        self
-    }
-
-    /// Generate a proof; see [`runtime.prove()`][GenericRuntime::prove].
-    pub fn prove(self) -> Result<Proof> {
-        self.runtime.prove(
-            self.program,
-            self.private_inputs,
-            self.public_inputs,
-            self.constant_inputs,
-        )
-    }
-
-    /// Generate a proof with parameters for that proof system; see
-    /// [`runtime.prove_with_parameters()`][GenericRuntime::prove_with_parameters].
-    pub fn prove_with_parameters(
-        self,
-        parameters: &B::ProverParameters,
-        transcript: &mut Transcript,
-    ) -> Result<Proof> {
-        self.runtime.prove_with_parameters(
-            self.program,
-            self.private_inputs,
-            self.public_inputs,
-            self.constant_inputs,
-            parameters,
-            transcript,
-        )
-    }
-}
-
-/// A builder for verifying a proof.
-///
-/// This is offered as a convenience for building the arguments necessary for the
-/// [`verify`][GenericRuntime::verify] function.
-pub struct VerificationBuilder<'r, 'p, 'a, T: marker::Zkp, B: ZkpBackend> {
-    runtime: &'r GenericRuntime<T, B>,
-    program: &'p CompiledZkpProgram,
-    proof: Option<&'a Proof>,
-    constant_inputs: Vec<ZkpProgramInput>,
-    public_inputs: Vec<ZkpProgramInput>,
-}
-
-impl<'r, 'p, 'a, T: marker::Zkp, B: ZkpBackend> VerificationBuilder<'r, 'p, 'a, T, B> {
-    /// Create a new `VerificationBuilder`. It's typically more convenient to create a
-    /// verification builder via
-    /// [`runtime.verification_builder()`][GenericRuntime::verification_builder].
-    pub fn new(runtime: &'r GenericRuntime<T, B>, program: &'p CompiledZkpProgram) -> Self
-    where
-        T: marker::Zkp,
-        B: ZkpBackend,
-    {
-        Self {
-            runtime,
-            program,
-            proof: None,
-            public_inputs: vec![],
-            constant_inputs: vec![],
-        }
-    }
-
-    /// Add the proof to verify.
-    pub fn proof(mut self, proof: &'a Proof) -> Self {
-        self.proof = Some(proof);
-        self
-    }
-
-    /// Add a constant input to the verification builder.
-    pub fn constant_input(mut self, input: impl Into<ZkpProgramInput>) -> Self {
-        self.constant_inputs.push(input.into());
-        self
-    }
-
-    /// Add multiple constant inputs to the verification builder.
-    pub fn constant_inputs<I>(mut self, inputs: I) -> Self
-    where
-        I: IntoIterator<Item = T>,
-        ZkpProgramInput: From<T>,
-    {
-        self.constant_inputs
-            .extend(inputs.into_iter().map(ZkpProgramInput::from));
-        self
-    }
-
-    /// Add a public input to the verification builder.
-    pub fn public_input(mut self, input: impl Into<ZkpProgramInput>) -> Self {
-        self.public_inputs.push(input.into());
-        self
-    }
-
-    /// Add multiple public inputs to the verification builder.
-    pub fn public_inputs<I>(mut self, inputs: I) -> Self
-    where
-        I: IntoIterator<Item = T>,
-        ZkpProgramInput: From<T>,
-    {
-        self.public_inputs
-            .extend(inputs.into_iter().map(ZkpProgramInput::from));
-        self
-    }
-
-    /// Verify that `self.proof` satisfies `self.program`; see
-    /// [`runtime.verify()`][GenericRuntime::verify].
-    ///
-    /// # Remarks
-    /// Will error if the underlying `verify` call errors, or if a proof has not yet been
-    /// supplied to the builder. That is, you must call [`Self::proof`] before calling this
-    /// function.
-    pub fn verify(self) -> Result<()> {
-        let proof = self.proof.ok_or_else(|| {
-            Error::zkp_builder_error(
-                "You must supply a proof to the verification builder before calling `verify`",
-            )
-        })?;
-        self.runtime.verify(
-            self.program,
-            proof,
-            self.public_inputs,
-            self.constant_inputs,
-        )
-    }
-}
